+++ src/components/Table/utils/TableState.ts
@@ -2,9 +2,9 @@ import { camelCase } from "change-case";
 import { makeAutoObservable, observable, ObservableSet, reaction } from "mobx";
 import React from "react";
 import { GridDataRow } from "src/components/Table/components/Row";
-import { RowStates } from "src/components/Table/components/RowStates";
 import { GridSortConfig } from "src/components/Table/GridTable";
 import { Direction, GridColumnWithId } from "src/components/Table/types";
+import { RowStates } from "src/components/Table/utils/RowStates";
 import { ASC, DESC, HEADER, KEPT_GROUP, reservedRowKinds } from "src/components/Table/utils/utils";
 import { isFunction } from "src/utils";
 import { assignDefaultColumnIds } from "./columns";
